Refactor example code to utilize helper functions for improved readability and maintainability
