+++ yy_cybergear/example/exmp_02_operation_sin_wave.cpp
@@ -44,141 +44,20 @@
 #include "yy_cybergear/protocol_types.hpp"
 #include "yy_socket_can/can_runtime.hpp"
 
+#include "exmp_helper.hpp"
+
 namespace
 {
 std::atomic<bool> g_running{true};
 void handle_sigint(int) { g_running = false; }
 
-// Aliases used across helpers (match exmp_05)
-using Clock = std::chrono::steady_clock;
-
-// Centralized logging lives in yy_cybergear/logging.hpp
-inline void print_status(const yy_cybergear::CyberGear & cg, double t_sec)
-{
-  std::cout << yy_cybergear::logging::formatStatusLine(cg, t_sec) << '\n';
-}
-
-inline void print_params(const yy_cybergear::CyberGear & cg)
-{
-  std::cout << yy_cybergear::logging::formatParamsSummary(cg);
-}
-
-// Register a wide handler and dispatch frames into all CyberGear mirrors
-inline void register_can_handler(
-  yy_socket_can::CanRuntime & rt, std::vector<yy_cybergear::CyberGear> & cgs, bool verbose)
-{
-  const uint32_t min_id = 0u;
-  const uint32_t max_id = CAN_EFF_MASK;
-  rt.registerHandler(min_id, max_id, [verbose, &cgs](const struct can_frame & f) {
-    for (auto & cg : cgs) (void)cg.dispatchAndUpdate(f);
-    if (verbose) {
-      const uint32_t id = f.can_id & CAN_EFF_MASK;
-      std::cout << "RX 0x" << std::hex << std::uppercase << id << std::dec
-                << " dlc=" << int(f.can_dlc) << "\n";
-    }
-  });
-}
-
-// Preflight: issue initial requests and wait until all are ready (with retries)
-inline void preflight_sync(
-  yy_socket_can::CanRuntime & rt, const std::string & ifname,
-  std::vector<yy_cybergear::CyberGear> & cgs, const std::vector<uint16_t> & preflight_params,
-  bool verbose)
-{
-  // Initial issue of ReadParam and GetDeviceId
-  for (std::size_t i = 0; i < cgs.size(); ++i) {
-    for (uint16_t idx : preflight_params) {
-      struct can_frame tx
-      {
-      };
-      cgs[i].buildReadParam(idx, tx);
-      rt.post(yy_socket_can::TxRequest{ifname, tx});
-      if (verbose) {
-        const uint32_t id = tx.can_id & CAN_EFF_MASK;
-        std::cout << "TX 0x" << std::hex << std::uppercase << id << std::dec << " (ReadParam 0x"
-                  << std::hex << std::uppercase << idx << std::dec << ")\n";
-      }
-    }
-    struct can_frame tx
-    {
-    };
-    cgs[i].buildGetDeviceId(tx);
-    rt.post(yy_socket_can::TxRequest{ifname, tx});
-    if (verbose) {
-      const uint32_t id = tx.can_id & CAN_EFF_MASK;
-      std::cout << "TX 0x" << std::hex << std::uppercase << id << std::dec << " (GetDeviceId)\n";
-    }
-  }
-
-  // Wait for completion with periodic retries
-  bool all_ready = false;
-  auto last_retry = Clock::now();
-  while (g_running && !all_ready) {
-    all_ready = true;
-    for (const auto & cg : cgs) {
-      if (!cg.isInitializedFor(preflight_params, /*require_uid=*/true)) {
-        all_ready = false;
-        break;
-      }
-    }
-    if (all_ready) break;
-
-    const auto now = Clock::now();
-    if (now - last_retry >= std::chrono::milliseconds(200)) {
-      last_retry = now;
-      for (std::size_t i = 0; i < cgs.size(); ++i) {
-        const auto & cg = cgs[i];
-        for (uint16_t idx : preflight_params) {
-          if (!cg.isParamInitialized(idx)) {
-            struct can_frame tx
-            {
-            };
-            cgs[i].buildReadParam(idx, tx);
-            rt.post(yy_socket_can::TxRequest{ifname, tx});
-            if (verbose) {
-              const uint32_t id = tx.can_id & CAN_EFF_MASK;
-              std::cout << "RETRY TX 0x" << std::hex << std::uppercase << id << std::dec
-                        << " (ReadParam 0x" << std::hex << std::uppercase << idx << std::dec
-                        << ")\n";
-            }
-          }
-        }
-        if (!cg.isUidInitialized()) {
-          struct can_frame tx
-          {
-          };
-          cgs[i].buildGetDeviceId(tx);
-          rt.post(yy_socket_can::TxRequest{ifname, tx});
-          if (verbose) {
-            const uint32_t id = tx.can_id & CAN_EFF_MASK;
-            std::cout << "RETRY TX 0x" << std::hex << std::uppercase << id << std::dec
-                      << " (GetDeviceId)\n";
-          }
-        }
-      }
-    }
-    std::this_thread::sleep_for(std::chrono::milliseconds(20));
-  }
-}
-
-// Wait until user presses Enter (or SIGINT occurs). Returns true if Enter was pressed.
-inline bool wait_for_enter_or_sigint()
-{
-  while (g_running) {
-    struct pollfd pfd;
-    pfd.fd = 0;  // stdin
-    pfd.events = POLLIN;
-    pfd.revents = 0;
-    const int pret = ::poll(&pfd, 1, 200);  // 200 ms timeout
-    if (pret > 0 && (pfd.revents & POLLIN)) {
-      char ch = 0;
-      const ssize_t n = ::read(0, &ch, 1);
-      if (n > 0 && (ch == '\n' || ch == '\r')) return true;  // start on Enter
-      // If other characters were typed, keep polling until newline arrives
-    }
-  }
-  return false;
-}
+using exmp_helper::check_for_errors;
+using exmp_helper::preflight_sync;
+using exmp_helper::print_params;
+using exmp_helper::print_status;
+using exmp_helper::register_can_handler;
+using exmp_helper::wait_for_enter_or_sigint;
+using exmp_helper::Clock;
 
 }  // namespace
 
@@ -315,7 +194,16 @@ int main(int argc, char ** argv)
     yy_cybergear::SPEED_KP,
     yy_cybergear::SPEED_KI,
   };
-  preflight_sync(rt, ifname, cgs, preflight_params, verbose);
+  preflight_sync(g_running, rt, ifname, cgs, preflight_params, verbose);
+
+  // Check for errors after preflight sync
+  for (const auto & cg : cgs) {
+    if (check_for_errors(cg)) {
+      std::cerr << "ERROR: Motor faults detected during initialization. Exiting.\n";
+      rt.stop();
+      return EXIT_FAILURE;
+    }
+  }
 
   // Print collected parameters per motor
   std::cout << "\nCollected parameters (including UID, excluding Status):\n";
@@ -324,7 +212,7 @@ int main(int argc, char ** argv)
   }
 
   std::cout << "\nPress Enter to start control (Ctrl+C to exit) ..." << std::endl;
-  if (!wait_for_enter_or_sigint()) {
+  if (!wait_for_enter_or_sigint(g_running)) {
     rt.stop();
     return EXIT_SUCCESS;
   }
@@ -363,6 +251,16 @@ int main(int argc, char ** argv)
       const auto start = Clock::now();
       const auto deadline = start + dt_ns;
       const double t_now = std::chrono::duration<double>(start - t0).count();
+
+      // Check for errors in all motors and exit if any are found
+      for (const auto & cg : cgs) {
+        if (check_for_errors(cg)) {
+          std::cerr << "ERROR: Stopping all motors due to detected faults.\n";
+          g_running = false;
+          break;
+        }
+      }
+      if (!g_running) break;
 
       // Print snapshot for each motor
       for (const auto & cg : cgs) {
@@ -410,5 +308,6 @@ int main(int argc, char ** argv)
   }
 
   rt.stop();
-  return EXIT_SUCCESS;
+  // Exit with failure if we stopped due to faults
+  return g_running ? EXIT_SUCCESS : EXIT_FAILURE;
 }
